# openapi.yaml
openapi: 3.0.1
info:
  title: CMS API
  description: API for Content Management System
  version: 1.0.0
paths:
...
  /contents/{contentId}:
    get:
      summary: Get content by ID
      parameters:
        - name: contentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successfully retrieved content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentResponseDTO'
    put:
      summary: Update content by ID
      parameters:
        - name: contentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentRequestDTO'
      responses:
        '200':
          description: Successfully updated content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentResponseDTO'
    delete:
      summary: Delete content by ID
      parameters:
        - name: contentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Successfully deleted content
  /content-histories:
    post:
      summary: Create new content history
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentHistoryRequestDTO'
      responses:
        '201':
          description: Successfully created content history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentHistoryResponseDTO'
    get:
      summary: Get all content histories
      responses:
        '200':
          description: Successfully retrieved content histories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContentHistoryResponseDTO'
  /content-histories/{changeId}:
    get:
      summary: Get content history by ID
      parameters:
        - name: changeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successfully retrieved content history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentHistoryResponseDTO'
    put:
      summary: Update content history by ID
      parameters:
        - name: changeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentHistoryRequestDTO'
      responses:
        '200':
          description: Successfully updated content history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentHistoryResponseDTO'
    delete:
      summary: Delete content history by ID
      parameters:
        - name: changeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Successfully deleted content history
  /content-histories/content/{contextId}:
    get:
      summary: Get content history by content ID
      parameters:
        - name: contextId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successfully retrieved content histories by content ID
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContentHistoryResponseDTO'
  /content-histories/user/{userId}:
    get:
      summary: Get content history by user ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successfully retrieved content histories by user ID
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContentHistoryResponseDTO'
...
  /reviews/content/{contentId}:
    get:
      summary: Get reviews by content ID
      parameters:
        - name: contentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successfully retrieved reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReviewResponseDTO'
...
components:
  schemas:
    ContentResponseDTO:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        body:
          type: string
        status:
          $ref: '#/components/schemas/ContentStatus'
        createdBy:
          $ref: '#/components/schemas/UserResponseDTO'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ContentRequestDTO:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
        status:
          $ref: '#/components/schemas/ContentStatus'
    ContentHistoryResponseDTO:
      type: object
      properties:
        id:
          type: integer
        contentId:
          type: integer
        userId:
          type: integer
        changeDescription:
          type: string
        createdAt:
          type: string
          format: date-time
    ContentHistoryRequestDTO:
      type: object
      properties:
        contentId:
          type: integer
        userId:
          type: integer
        changeDescription:
          type: string
...
    ReviewResponseDTO:
      type: object
      properties:
        id:
          type: integer
        contentId:
          type: integer
        userId:
          type: integer
        rating:
          type: integer
        comment:
          type: string
        status:
          $ref: '#/components/schemas/ReviewStatus'
...
    UserResponseDTO:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
...
    Role:
      type: string
      enum: [REVIEWER, CONTRIBUTOR, ADMINISTRATOR]
    ContentStatus:
      type: string
      enum: [DRAFT, IN_REVIEW, NEEDS_EDITING, APPROVED, COMPLETED, PUBLISHED]
    ReviewStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED]
...